DROTH-615 (US-418) Create New Fetch to return All Active Huoltotie Assets inside a given Polygon Area: -Unit test and some code improvements

+++ digiroad2-api-oth/src/main/scala/fi/liikennevirasto/digiroad2/MaintenanceRoadApi.scala
@@ -38,7 +38,7 @@ class MaintenanceRoadApi(val linearAssetService: LinearAssetService) extends Sca
 
   get("/huoltotiet/:areaId"){
     var areaId = params("areaId")
-    val maintenanceAsset = linearAssetService.getActiveMaintenanceRoadByPolygon(areaId, typeId)
+    val maintenanceAsset = linearAssetService.getActiveMaintenanceRoadByPolygon(areaId.toInt, typeId)
     createJson(maintenanceAsset)
   }
